Move NpmImporter to pipeline directory
