+++ vulnerabilities/pipelines/npm_importer.py
