+++ vulnerabilities/tests/test_data/npm/npm_sample.json
