+++ vulnerabilities/tests/test_data/npm/parse-advisory-npm-expected.json
