+++ vulnerabilities/tests/test_data_source.py
@@ -25,10 +25,10 @@ from vulnerabilities.importers.github_osv import GithubOSVImporter
 from vulnerabilities.importers.gitlab import GitLabAPIImporter
 from vulnerabilities.importers.istio import IstioImporter
 from vulnerabilities.importers.mozilla import MozillaImporter
-from vulnerabilities.importers.npm import NpmImporter
 from vulnerabilities.importers.retiredotnet import RetireDotnetImporter
 from vulnerabilities.importers.ruby import RubyImporter
 from vulnerabilities.oval_parser import OvalParser
+from vulnerabilities.pipelines.npm_importer import NpmImporter
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 TEST_DATA = os.path.join(BASE_DIR, "test_data/")
